fix: rename functions.d.ts to types.d.ts
